requirements.txt
# Updated requirements.txt for Railway deployment
flask==2.2.3
werkzeug==2.2.3
supabase==1.0.3
python-dotenv==1.0.0
requests==2.28.2
torch==2.6.0
torchaudio==2.6.0
huggingface_hub==0.28.1
transformers==4.49.0
soundfile==0.12.1
safetensors==0.5.3
pillow
numpy
tokenizers==0.21.0
openai==1.2.0
elevenlabslib>=0.32.0
gunicorn==21.2.0
boto3==1.34.0